refactor: renamed files to be inline with file naming convention
