+++ lua/smart-motion/visualizers/pass_through.lua
